Create codesign-cogstat-intel-python3_11.sh (#253)
This is an updated version (from py3.8 to py3.11) of the codesign bash I've written, so the app can be signed (which means it won't throw security errors when running -- remember the issue we had with a peer reviewer who refused to open it because of the security notification on mac). I've replaced my signing identity for privacy reasons but this script will help anyone with a valid apple dev account to sign the app. The mac compilation instructions will be updated as well.